Relocate userSlice into users folder
