+++ package.json
@@ -59,9 +59,7 @@
       "!src/index.tsx",
       "!src/react-app-env.d.ts",
       "!src/setupTests.ts",
-      "!src/reportWebVitals.ts",
-      "!src/store/hooks.ts",
-      "!src/store/index.ts"
+      "!src/reportWebVitals.ts"
     ]
   }
 }
