+++ sonar-project.properties
@@ -14,6 +14,6 @@ sonar.sources=./src
 sonar.sourceEncoding=UTF-8
 
 sonar.test.inclusions=src/**/*.test.ts, src/**/*.test.tsx
-sonar.coverage.exclusions=src/index.tsx, src/setupTests.ts, src/reportWebVitals.ts, src/store/hooks.ts, src/store/index.ts
+sonar.coverage.exclusions=src/index.tsx, src/setupTests.ts, src/reportWebVitals.ts,
 
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
